my-ai-app/backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
requests
python-dotenv
sse-starlette
